feat: Add Integration Id plugin
Add the AddIntegrationIdAction plugin that links a profile with its corresponding identity in an external system. The plugin allows users to specify the external ID and the name of the external system. Additional data related to the external system can also be added. The plugin saves the external system ID in the profile's metadata.system.integrations, where the key is the system name and the value is the ID and data.

Plugin metadata:
- Name: Add Integration Id
- Description: Link profile with their corresponding identities in an external system.
- Group: Operations
- Purpose: Collection, Segmentation

Inputs:
- payload: This port takes a payload object.

Outputs:
- payload: This port returns a payload object.
- error: This port returns an error message if the plugin fails.

Version: 0.8.2
License: MIT + CC
Author: Risto Kowaczewski

Form:
- External ID: Reference external ID.
-