feat(action): add description for External System Name field
This commit adds a description for the "External System Name" field in the `add_integration_id_action.py` file. The description states that the name will be lower-cased and spaces will be replaced by hyphens.

The description is necessary to provide clarity and guidance to users when filling out the form.