feat: add support for workflow tags in metadata and endpoints
This commit adds support for workflow tags in the metadata and endpoints.

In the `FlowMetaData` class, the `projects` attribute has been renamed to `tags`. The default value for `tags` has been changed to ["General"].

In the `Flow` class, the `projects` attribute has been renamed to `tags`. The default value for `tags` has been changed to ["General"].

In the `FlowRecord` class, the `projects` attribute has been renamed to `tags`.

In the `TableService` class, the `projects` parameter has been changed to `tags` in the `_select_in_deployment_mode` method.

In the `WorkflowService` class, a new method `load_in_current_context` has been added to load a workflow by its ID.

In the `flow_endpoint.py` file, a new function `_record_exists_in_current_context` has been added to check if a