feat(database): Add 'file_name' column to 'table' table
This commit adds a new column called 'file_name' to the 'table' table in the MySQL database schema. This column is of type String with a maximum length of 64 characters. The purpose of this column is to store the name of the file associated with the table.

The 'file_name' column is also added to the 'Flow' domain class. This allows for the storage and retrieval of the file name associated with a flow. The 'file_name' property is optional and can be set to None if no file is associated with the flow.

The 'file_name' column is also included in the 'workflow_mapping' module. The column is mapped to the 'flow_record' and 'workflow_table' objects, allowing for the proper storage and retrieval of the file name in the MySQL database.

This enhancement improves the functionality of the Tracardi system by providing a way to associate a file name with a table