feat(storage): Implement cache invalidation in various services #1.2.x
Cache invalidation has been implemented in the 'insert' and 'delete' methods of several services. These include:

EventSourceService, EventReshapingService, EventValidationService, EventMappingService, EventToProfileMappingService, ConsentDataComplianceService in both 'tracardi' and 'com_tracardi' repositories.

This feature ensures that the cache is updated immediately when a record is inserted or deleted, keeping the cache data consistent with the data in the database. This enhances the performance and reliability of the system.