feat: Add server state management and update packages (#1.2.x)
This commit introduces server state management with the 'server_ready' flag
in a new state.py. This flag is used to monitor the server state and is
updated in startup.py when the app starts successfully.

The health_endpoint.py is refactored to use the 'server_ready' flag for
health checks, responding with 'pong' if the server is ready or raising a
404 HTTP exception with the detail "Not ready" otherwise.

A 'context' is added in health_endpoint.py to provide additional context
data when the server is ready.

The 'dotdict_parser' package is updated from version 0.6.1 to 0.7.1 in
com_tracardi/requirements.txt, setup.py, and tracardi/requirements.txt.

A new unit test for the 'dotdict_parser' flatten function is added in
test/unit/test_flatten.py.